+++ spec/rails_app/app/controllers/posts_controller.rb
@@ -87,7 +87,7 @@ class PostsController < ApplicationController
     @post.destroy
 
     respond_to do |format|
-      format.html { redirect_to posts_url }
+      format.html { redirect_to posts_url, :notice => "Post destroyed." }
       format.json { head :no_content }
     end
   end